+++ python/oneflow/utils/__init__.py
@@ -14,5 +14,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 from oneflow.framework.config_util import api_load_library_now as load_library
-from oneflow.utils.torch.from_torch import from_torch
-from oneflow.utils.torch.to_torch import to_torch
+from oneflow.utils.torch.from_or_to_torch_tensor import from_torch, to_torch
